+++ widgets/Gruntfile.js
@@ -25,7 +25,7 @@ module.exports = function (grunt) {
                 sourceMap: false
             },
             files: {
-                'build/css/nfhs.css': 'src/less/nfhs.less',
+                'build/css/digitalscout.css': 'src/less/digitalscout.less',
                 'build/usa_today.css': 'examples/usa_today.less',
                 'build/css/base.css': 'src/less/base/base.less'
             }