+++ widgets/src/js/directives/footballDriveChart.js
@@ -4,21 +4,31 @@ angular.module('hsgc')
       restrict: 'EA',
       templateUrl: 'templates/footballDriveChart.html',
       link: function(scope, element) {
+        // configuration
+        var fontTeams = "Graduate, Open Sans, sans-serif";
+        var fontYards = "Open Sans, sans-serif";
+        var yardLineWidth = 3;
+        var maxEndzoneTextCharacters = 22;
+        var scrimmageLineColor = "#ffff00";
+
+        // internal tracking variables
         scope.selectedPeriod = 0;
         scope.lastPlay = {};
         scope.previousPlay = {};
         scope.hasPreviousPlay = false;
-        var paper = typeof(window.paper) === 'undefined' ? null : window.paper;
-        var firstload = true;
-        var footballFieldCanvas = null;
-        var fontTeams = "Graduate, sans-serif";
-        var fontYards = "Open Sans, sans-serif";
-        var yardLineWidth = 3;
-        var maxEndzoneTextCharacters = 22;
 
+        var footballFieldCanvas = null;
+        var footballFieldCanvasSize = { width: -1 };
+        // save reference to the paper instance, so multiple contexts may be used on one page
+        var paper = null, paperDriveGroup = null;
+        // yardlines and textHeight adjust per size, but need to be used by multiple functions separately
+        var yardLines = [], textHeight = 0;
+
+        var firstload = true;
+        
         scope.$watch('selectedDetailTab', function(newValue, oldValue) {
           if (newValue === 5 && newValue !== oldValue) {
-            scope.resize(true);
+            scope.resize(true, false);
           }
         });
 
@@ -34,7 +44,8 @@ angular.module('hsgc')
             footballFieldCanvas = $(element.find('canvas')[0]);
             if (footballFieldCanvas) {
               // if play-by-play not available, canvas won't be loaded; but when it is, set it all up:
-              paper.setup(footballFieldCanvas[0]);
+              paper = window.paper.setup(footballFieldCanvas[0]);
+              scope.drawField();
 
               window.angular.element($window).on('resize', scope.resize);
               scope.$on('$destroy', function() {
@@ -47,52 +58,58 @@ angular.module('hsgc')
           scope.previousPlay = scope.playByPlay[scope.playByPlay.length - 2];
           scope.hasPreviousPlay = scope.playByPlay.length > 2;
 
-          scope.resize(true);
+          scope.resize(false, true);
         });
 
-        scope.resize = function(forceRedraw) {
+        scope.resize = function(becomingVisible, newData) {
           // canvas won't be found if play-by-play not available
-          if (!footballFieldCanvas) {
+          if (!footballFieldCanvas || !paper) {
             $log.debug('canvas element for the drive chart not available; play-by-play probably not available, which is fine', element);
             return;
           }
             
           // calculate the drive chart size
-          var desiredWidth = element.parent().width() - 20;
-          var desiredHeight = 0;
+          var desiredWidth = element.width() - parseInt(element.css("border-left-width"), 10) - parseInt(element.css("border-right-width"), 10);
           if (desiredWidth < 1) {
-            desiredWidth = footballFieldCanvas.width();
+            // borders load regardless; but width only works if already visible, so parent.width() may have returned 0, try with the parent size (the container that is already visible but probably doesn't have the borders on it) minus borders of the canvas' parent (which may not be visible so can't use width, but can load borders for it)
+            var borders = parseInt(footballFieldCanvas.parent().css("border-left-width"), 10) - parseInt(footballFieldCanvas.parent().css("border-right-width"), 10);
+            desiredWidth = element.parent().width() - borders;
           }
-
           // football fields including endzones are 4:9 = height:width ratio
-          desiredHeight = Math.floor(desiredWidth * 4 / 9);
+          var desiredHeight = Math.floor(desiredWidth * 4 / 9);
 
-          // only redraw if the size actually changed
-          if (forceRedraw === true ||
-            (element.is(':visible') && footballFieldCanvas && (footballFieldCanvas.width() != desiredWidth || footballFieldCanvas.height() != desiredHeight))) {
+          // only redraw field if the size actually changed
+          if ((becomingVisible || element.is(':visible')) && footballFieldCanvas && footballFieldCanvasSize.width != desiredWidth) {
             $log.debug('Drive Chart size changed, redrawing', footballFieldCanvas.width(), footballFieldCanvas.height(), desiredWidth, desiredHeight);
             footballFieldCanvas.prop({
               width: desiredWidth,
               height: desiredHeight
             });
 
-            paper.view.viewSize = new paper.Size(desiredWidth, desiredHeight);
+            footballFieldCanvasSize = new paper.Size(desiredWidth, desiredHeight);
+            paper.view.viewSize = footballFieldCanvasSize;
+            scope.drawField();
             scope.drawDrive();
-          /*
+          } else if (newData === true) {
+            if (footballFieldCanvas) {
+              $log.debug('Drive Chart size not changed, but datacastLoaded so redrawing drive', footballFieldCanvas.width(), footballFieldCanvas.height(), desiredWidth, desiredHeight, element.is(':visible'), footballFieldCanvas.width() != desiredWidth);
+            }
+
+            scope.drawDrive();
           } else {
             $log.debug('window resized but field size constant; skipping redraw');
-          */
           }
-        };       
+        };
 
-        scope.drawDrive = function() {
+        scope.drawField = function() {
+          paper.project.clear();
           var fieldSize = paper.view.viewSize;
           var field = new paper.Path.Rectangle(new paper.Rectangle(new paper.Point(0,0), fieldSize));
           field.fillColor = "#008A2E";
 
           var endzoneWidth = fieldSize.width / 12;
           var playingFieldWidth = fieldSize.width - 2 * endzoneWidth;
-          var yardLines = [];
+          yardLines = [];
           for (var i = 0; i <= 100; i++) {
             yardLines.push(endzoneWidth + i * playingFieldWidth / 100);
           }
@@ -123,7 +140,7 @@ angular.module('hsgc')
           homeTeamName.fontFamily = fontTeams;
           homeTeamName.rotate(90);
 
-          var textHeight = (yardLines[5] - yardLines[0]) * 0.7;
+          textHeight = (yardLines[5] - yardLines[0]) * 0.7;
           var textSizes = {};
           var yardNumbers = [1, 2, 3, 4, 5, 0];
           for (i = 0; i < yardNumbers.length; i++) {
@@ -194,12 +211,20 @@ angular.module('hsgc')
               scope.drawHashMark(yardLines[i], bottomMiddleHashMarkY, false);
             }
           }
+        };
 
+        scope.drawDrive = function() {
+          if (paperDriveGroup) {
+            paperDriveGroup.remove();
+            paperDriveGroup = null;
+          }
+          
+          var fieldSize = paper.view.viewSize;
           var homeTeamIsOnOffense = scope.lastPlay.TeamSeasonId == scope.homeTeamSeasonId;
           var scrimmageYardLine = homeTeamIsOnOffense ? 100 - scope.lastPlay.Spot : scope.lastPlay.Spot;
           var scrimmageLineX = yardLines[scrimmageYardLine];
-          var scrimmageLine = new paper.Path.Line(new paper.Point(scrimmageLineX, 0), new paper.Point(scrimmageLineX, fieldSize.height));
-          scrimmageLine.strokeColor = "#ffff00";
+          var scrimmageLine = new paper.Path(new paper.Point(scrimmageLineX, 0), new paper.Point(scrimmageLineX, fieldSize.height));
+          scrimmageLine.strokeColor = scrimmageLineColor;
           scrimmageLine.strokeWidth = 5;
 
           var arrowOffset, arrowRotation, arrowColor;
@@ -228,6 +253,8 @@ angular.module('hsgc')
           var firstDownLine = new paper.Path.Line(new paper.Point(firstDownLineX, 0), new paper.Point(firstDownLineX, fieldSize.height));
           firstDownLine.strokeColor = "#00ffff";
           firstDownLine.strokeWidth = 5;
+
+          paperDriveGroup = new paper.Group([scrimmageLine, arrow, firstDownLine]);
 
           paper.view.draw();
         };
@@ -280,6 +307,8 @@ angular.module('hsgc')
             return "3rd";
           } else if (number == 4) {
             return "4th";
+          } else {
+            return "-";
           }
         };
       }