+++ widgets/src/js/directives/footballGameLeaders.js
@@ -2,6 +2,6 @@ angular.module('hsgc')
 	.directive('gameLeaders', function() {
 		return {
 			restrict: 'EA',
-			templateUrl: 'templates/gameLeaders.html'
+			templateUrl: 'templates/footballGameLeaders.html'
 		};
 	});